+++ src/Http/Controllers/ServerController.php
@@ -2,10 +2,10 @@
 
 namespace Eyewitness\Eye\Http\Controllers;
 
-use Illuminate\Routing\Controller;
+use Eyewitness\Eye\Http\Controllers\BaseController;
 use Eyewitness\Eye\Eye;
 
-class ServerController extends Controller
+class ServerController extends BaseController
 {
     /**
      * Run the server ping command.
@@ -16,5 +16,16 @@ class ServerController extends Controller
     public function ping(Eye $eye)
     {
         return response()->json($eye->runAllChecks(), 200);
+    }
+
+    /**
+     * Handle the old depreciated route to advise incoming API calls of the new version
+     * without causing any service interruptions.
+     *
+     * @return json
+     */
+    public function moved()
+    {
+        return $this->jsonp(['msg' => 'Moved'], 410);
     }
 }